shared-platform/pyproject.toml
[project]
name = "shared-platform"
version = "0.1.0"
description = "Reusable infrastructure layer for AI systems"
authors = [{ name = "Oren Salmi" }]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
